+++ src/__tests__/mocks.ts
@@ -45,6 +45,22 @@ export const ccpaDataExample = {
     noticeGiven: false
 }
 
+export const gppDataExample = {
+    gppString: "DBABMA~CPXxRfAPXxRfAAfKABENB-CgAAAAAAAAAAYgAAAAAAAA",
+    applicableSections: [7],
+    sections: {
+        usnat: {
+            Version: 1,
+            SharingNotice: 2,
+            SaleOptOut: 2
+        }
+    }
+}
+
+export const gppSectionChangePayloadExample = {
+    data: "usnat"
+}
+
 const ucLabels: UsercentricsLabels = {
     btnAcceptAll: "btnAcceptAll",
     btnDeny: "btnDeny",
@@ -268,7 +284,8 @@ const disclosure: ConsentDisclosure = {
 }
 
 const consentDisclosureObject: ConsentDisclosureObject = {
-    disclosures: [disclosure]
+    disclosures: [disclosure],
+    sdks: [],
 }
 
 const ucService: UsercentricsService = {